requirements.txt
aiofiles>=24.1.0
fastapi[standard]>=0.115.12
insightface>=0.7.3
lancedb>=0.24.0
loguru>=0.7.3
numpy>=1.26.0
onnx>=1.13.1
onnxruntime-gpu>=1.22.0
opencv-python>=4.11.0.86
pandas>=2.3.0
pydantic>=2.11.5
pydantic-settings>=2.9.1
python-dotenv>=1.1.0
pyyaml>=6.0.2
scipy>=1.15.3
streamlit>=1.45.1
typer>=0.16.0
sqlalchemy>=2.0.41
paho-mqtt